Added icinga exchange file. closes #1.
Licence set as LGPL to be compatible with the current version of
XenAPI.py as stored in this repo. As the check_xen.py file has no
licence set in the file, will assume the same as xen-api is released
under - see discussion below.

check_xen.py
------------

Initial file from:

* https://xenserver.org/partners/developing-products-for-xenserver/20-dev-hints/120-xs-pool-check-nagios.html
* https://github.com/xapi-project/xen-api/blob/f77e7288839255690826d89c1ce750f23554e45e/ocaml/idl/binding_sanity_checks/tutorial_examples/nagios_plugin.py

Which gives a date of at least 2009-06-06. Xenserver.org site has no
date, so cannot be used to date the project. The version in this repo
has more updates however.

Another version which contains only one line difference can be found
here:

* ftp://ftp.ajlc.waterloo.on.ca/pub/check_xen.py

Which has (at time of commit) a modification date of 2010-04-10.

The licence listed on the xapi-project/xen-api github repo is the LGPL,
and as this has the oldest date recorded then this will be the licence
for the file.

XenAPI.py
---------

Initial file from:

* https://github.com/xapi-project/xen-api/blob/f77e7288839255690826d89c1ce750f23554e45e/ocaml/idl/binding_sanity_checks/tutorial_examples/XenAPI.py

Minor updated version:

* https://github.com/xapi-project/xen-api/blob/b022d1f0062a4ef606818caf419febf7aed66078/ocaml/idl/binding_sanity_checks/tutorial_examples/XenAPI.py

Version with minor differences from same time in the same repo:

* https://github.com/xapi-project/xen-api/blob/f77e7288839255690826d89c1ce750f23554e45e/scripts/examples/python/XenAPI.py

Majorly updated version:

* https://github.com/xapi-project/xen-api/blob/master/scripts/examples/python/XenAPI.py

Change in Licence of one of the versions of the file:

* https://github.com/xapi-project/xen-api/commit/e1f035c563b93b004c039bd2b74e238397210039

As the version in this repository is the same as the version with LGPL,
we will stick with LGPL until the XenAPI data changes - at which point,
licences will be applied as they appear in the files, and the codebase
will be dual licenced.